Move tty detection to its own module
